requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai
langchain-community
pymysql
sqlalchemy
pymilvus
paddlepaddle
paddleocr
cryptography
requests